[Fix] INT32 overflow in implicit GEMM kernels at large dims

+++ torchsparse/backend/convolution/convolution_backward_wgrad_implicit_gemm_cuda.cu
@@ -17,7 +17,7 @@ __pack_half2(const half x, const half y)
 
 // conv_backward_cuda_m16n16k64_m16n16k64_m16n16k16_f16f16f32
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f16f16f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, half *__restrict__ A, half *__restrict__ B, int *__restrict__ out_in_map, half *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f16f16f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, half *__restrict__ A, half *__restrict__ B, int *__restrict__ out_in_map, half *__restrict__ C)
 {
   int j_factors1 = (N + 15) / 16 / 1;
   int blockIdx_x = 0;
@@ -58,7 +58,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f16f16f3
   if constexpr (K_ld_check)
   {
     int A_ld_start = ((threadIdx.y * 256 % 16) % K_tile_padded) + ((threadIdx.x * 8 % 16) % K_tile_padded) + ((blockIdx_y / j_factors1 * 16) % K_tile_padded);
-    int A_ld_amount = min(A_ld_start + 8, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 8), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 2);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -69,7 +69,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f16f16f3
   if constexpr (N_ld_check)
   {
     int B_ld_start = (blockIdx_y % j_factors1) * 16 + (threadIdx.x * 8) % 16;
-    int B_ld_amount = min(B_ld_start + 8, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 8), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 2);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -368,7 +368,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f16f16f3
 }
 
 // conv_backward_cuda_m32n64k64_m32n32k64_m16n16k16_f16f16f32
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_f16f16f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, half *__restrict__ A, half *__restrict__ B, int *__restrict__ out_in_map, half *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_f16f16f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, half *__restrict__ A, half *__restrict__ B, int *__restrict__ out_in_map, half *__restrict__ C)
 {
   int j_factors1 = N / 16 / 4;
   int blockIdx_x = 0;
@@ -700,7 +700,7 @@ __global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_f16f16f3
 
 // conv_backward_cuda_m16n16k64_m16n16k64_m16n16k16_tf32tf32f32
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_tf32tf32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_tf32tf32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
 {
   int j_factors1 = (N + 15) / 16 / 1;
   int blockIdx_x = 0;
@@ -741,7 +741,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_tf32tf32
   if constexpr (K_ld_check)
   {
     int A_ld_start = ((threadIdx.y * 256 % 16) % K_tile_padded) + ((threadIdx.x * 8 % 16) % K_tile_padded) + ((blockIdx_y / j_factors1 * 16) % K_tile_padded);
-    int A_ld_amount = min(A_ld_start + 8, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 8), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 4);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -753,7 +753,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_tf32tf32
   if constexpr (N_ld_check)
   {
     int B_ld_start = (blockIdx_y % j_factors1) * 16 + (threadIdx.x * 8) % 16;
-    int B_ld_amount = min(B_ld_start + 8, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 8), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 4);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -969,7 +969,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_tf32tf32
 }
 
 // conv_backward_cuda_m32n64k64_m32n32k64_m16n16k16_tf32tf32f32
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_tf32tf32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_tf32tf32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
 {
   int j_factors1 = N / 16 / 4;
   int blockIdx_x = 0;
@@ -1219,7 +1219,7 @@ __global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_tf32tf32
 
 // conv_backward_cuda_m16n16k64_f32f32f32
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f32f32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f32f32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
 {
 
   int j_factors1 = (N + 15) / 16; 
@@ -1276,7 +1276,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f32f32f3
   if constexpr (K_ld_check) // IC % cta_M != 0 
   {
     int A_ld_start = channel_offset;
-    int A_ld_amount = min(A_ld_start + 4, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 4), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 4);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -1288,7 +1288,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f32f32f3
   if constexpr (N_ld_check) // OC % cta_N != 0
   {
     int B_ld_start = (blockIdx_n * 16) + ((threadIdx_x * 4) % 16); 
-    int B_ld_amount = min(B_ld_start + 4, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 4), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 4);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -1441,7 +1441,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode0_f32f32f3
 }
 
 // conv_backward_cuda_m32n64k64_f32f32f32
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_f32f32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode0_f32f32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, float *__restrict__ A, float *__restrict__ B, int *__restrict__ out_in_map, float *__restrict__ C)
 {
 
   int j_factors1 = (N + 63) / 64;
