+++ torchsparse/backend/convolution/convolution_backward_wgrad_implicit_gemm_sorted_cuda.cu
@@ -18,7 +18,7 @@ __pack_half2(const half x, const half y)
 
 // conv_backward_cuda_m16n16k64_m16n16k64_m16n16k16_f16f16f32
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f16f16f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, half *__restrict__ A, half *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, half *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f16f16f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, half *__restrict__ A, half *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, half *__restrict__ C)
 {
   const int K_tile = 16;
   int K_tile_padded = K_tile * ((K_original + K_tile - 1) / K_tile);
@@ -70,7 +70,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f16f16f3
   if constexpr (K_ld_check)
   {
     int A_ld_start = ((threadIdx.y * 256 % 16) % K_tile_padded) + ((threadIdx.x * 8 % 16) % K_tile_padded) + ((blockIdx_y / j_factors1 * 16) % K_tile_padded);
-    int A_ld_amount = min(A_ld_start + 8, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 8), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 2);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -81,7 +81,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f16f16f3
   if constexpr (N_ld_check)
   {
     int B_ld_start = (blockIdx_y % j_factors1) * 16 + (threadIdx.x * 8) % 16;
-    int B_ld_amount = min(B_ld_start + 8, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 8), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 2);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -383,7 +383,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f16f16f3
 
 
 // conv_backward_cuda_m32n64k64_m32n32k64_m16n16k16_f16f16f32
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_f16f16f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, half *__restrict__ A, half *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, half *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_f16f16f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, half *__restrict__ A, half *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, half *__restrict__ C)
 {
   float C_warp[32];
   __shared__ half A_shared[2560];
@@ -752,7 +752,7 @@ __global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_f16f16f3
 
 // conv_backward_cuda_m16n16k64_m16n16k64_m16n16k16_tf32tf32f32
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_tf32tf32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_tf32tf32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
 {
   const int K_tile = 16;
   int K_tile_padded = K_tile * ((K_original + K_tile - 1) / K_tile);
@@ -804,7 +804,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_tf32tf32
   if constexpr (K_ld_check)
   {
     int A_ld_start = ((threadIdx.y * 256 % 16) % K_tile_padded) + ((threadIdx.x * 8 % 16) % K_tile_padded) + ((blockIdx_y / j_factors1 * 16) % K_tile_padded);
-    int A_ld_amount = min(A_ld_start + 8, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 8), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 4);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -816,7 +816,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_tf32tf32
   if constexpr (N_ld_check)
   {
     int B_ld_start = (blockIdx_y % j_factors1) * 16 + (threadIdx.x * 8) % 16;
-    int B_ld_amount = min(B_ld_start + 8, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 8), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 4);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -1037,7 +1037,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_tf32tf32
 
 
 // conv_backward_cuda_m32n64k64_m32n32k64_m16n16k16_tf32tf32f32
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_tf32tf32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_tf32tf32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
 {
   float C_warp[32];
   __shared__ float A_shared[2560];
@@ -1324,7 +1324,7 @@ __global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_tf32tf32
 
 // conv_backward_cuda_m16n16k64_f32f32f32_sort
 template <int K_ld_factor, int N_ld_factor, bool K_ld_check, bool N_ld_check>
-__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f32f32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
+__global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f32f32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
 {
 
   int j_factors1 = (N + 15) / 16; 
@@ -1386,7 +1386,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f32f32f3
   if constexpr (K_ld_check) // IC % cta_M != 0 
   {
     int A_ld_start = channel_offset;
-    int A_ld_amount = min(A_ld_start + 4, K_original) - A_ld_start;
+    int A_ld_amount = min((int64_t)(A_ld_start + 4), K_original) - A_ld_start;
     int A_ld_bound = A_ld_amount / (K_ld_factor / 4);
 
     for (int i = 0; i < A_ld_bound; i++)
@@ -1398,7 +1398,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f32f32f3
   if constexpr (N_ld_check) // OC % cta_N != 0
   {
     int B_ld_start = (blockIdx_n * 16) + ((threadIdx_x * 4) % 16); 
-    int B_ld_amount = min(B_ld_start + 4, N) - B_ld_start;
+    int B_ld_amount = min((int64_t)(B_ld_start + 4), N) - B_ld_start;
     int B_ld_bound = B_ld_amount / (N_ld_factor / 4);
 
     for (int i = 0; i < B_ld_bound; i++)
@@ -1559,7 +1559,7 @@ __global__ void __launch_bounds__(32) conv_backward_cuda_setting1_mode1_f32f32f3
 
 
 // conv_backward_cuda_m32n64k64_f32f32f32_sort
-__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_f32f32f32(int M_fwd, int K_original, int N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
+__global__ void __launch_bounds__(64) conv_backward_cuda_setting2_mode1_f32f32f32(int M_fwd, int64_t K_original, int64_t N, int kernel_volume, int split_k_iters, int split_mask_len, int reduced_mask_len, int reorder_loc_len, float *__restrict__ A, float *__restrict__ B, int *__restrict__ reduced_mask, int *__restrict__ out_in_map, int *__restrict__ reorder_loc, float *__restrict__ C)
 {
 
   int j_factors1 = (N + 63) / 64;
